Remove bidsutils; fix bids editor main window
